feat(api/controller, api/middlewares): add controllers get and post users, cards and roles
